Fixing potential problems identified by FindBugs

+++ projects/OG-Analytics/src/com/opengamma/analytics/financial/equity/variance/VarianceSwapSensitivityCalculator.java
@@ -157,7 +157,7 @@ public final class VarianceSwapSensitivityCalculator extends DerivativeSensitivi
     final int nObsSoFar = swap.getObservations().length;
     final int nObsDidntHappen = swap.getObsDisrupted();
 
-    return (nObsExpected - nObsSoFar - nObsDidntHappen) / nObsExpected * swap.getVarNotional();
+    return (nObsExpected - nObsSoFar - nObsDidntHappen) / (double) nObsExpected * swap.getVarNotional();
   }
 
 }
